requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
addict=2.4.0=pypi_0
altair=5.3.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
blinker=1.8.2=pypi_0
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.2.2=hbcca054_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.3=pypi_0
fonttools=4.50.0=pypi_0
fsspec=2024.3.1=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
grpcio=1.62.1=pypi_0
huggingface-hub=0.22.2=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.1.0=pypi_0
iopath=0.1.10=pypi_0
ipykernel=6.29.3=pyhd33586a_0
ipynb-py-convert=0.4.6=pypi_0
ipython=8.22.2=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=py310hff52083_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multiscaledeformableattention=1.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.99=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.13=h7f8727e_0
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.1=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=pypi_0
pip=23.3.1=pypi_0
platformdirs=4.2.0=pyhd8ed1ab_0
portalocker=2.8.2=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=16.1.0=pypi_0
pycocotools=2.0.7=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.1.2=pypi_0
python=3.10.14=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.10=2_cp310
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.31.0=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
safetensors=0.4.2=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
setuptools=68.2.2=pypi_0
simpleitk=2.3.1=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.35.0=pypi_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.3.0=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.4.0=pypi_0
thop=0.1.1-2209072238=pypi_0
timm=0.9.16=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.1=pypi_0
torch=2.2.2=pypi_0
torchaudio=2.2.2=pypi_0
torchvision=0.17.2=pypi_0
tornado=6.1=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pyhd8ed1ab_0
triton=2.2.0=pypi_0
typing-extensions=4.10.0=pypi_0
typing_extensions=4.11.0=pyha770c72_0
tzdata=2024.1=pypi_0
ultralytics=8.2.9=pypi_0
urllib3=2.2.1=pypi_0
watchdog=4.0.1=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.2=pypi_0
wheel=0.41.2=pypi_0
xz=5.4.6=h5eee18b_0
yacs=0.1.8=pypi_0
yapf=0.40.2=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
